Adds a new model default apply strategy.
We need an apply strategy that always forces certain model defaults onto
the model config regardless of what config has set. We need this to
force through a models uuid and name on to the config.

Over time and the transition to Dqlite we can slowly look at moving
this.